feat(yt-dlp): download playlist within its own dir
Make a directory of that playlist title inside `~/videos`, and if it
doesn't exist (for example, we're downloading a video and not a
playlist) it will not create a new directory, and will put the video in
`~/videos`.

Special Thanks:
discord.com/channels/807245652072857610/807247901981016094/1257733938214994041

TODO: Create preset for different use cases using `--alias`.