chore(nvim): always overwrite chezmoi conflict
Conflict occurs when the original file inside `~/.config` is changed,
and we also change the chezmoi configuration file
`~/.local/share/chezmoi/dot_config`.

This will prompt what're you going to do with the conflict with a tty,
however `vim.fn.system` can't prompt the user for tty input.

Probably setting it to be `--force` is the best option, because we are
guaranteed that our chezmoi config is properly synced with our actual
config directory