feat(nvim): add some window and tab navigation
Assume `xx` as `<leader>xx`.
Well firstly, I renamed "split" into "window" as it is the correct
name.

And then, change `wx` and `tx` to be `wq` and `tq`. To make it the same
as the equivalent <C-w> keybind. And for `wx` specifically, it's for
swap split now.

Then, `Tab` and `Shift+Tab` now go to next/previous tab for easier
navigation.

Lastly, I add `w` + `hjkl` to move window to far (top/left/right/bottom)