refactor(nvim): add score to all codec on trackselect
This basically achieves more-or-less the same result as the previous
version, but better (in terms of code).

So this will add score to all codec. But the better the codec, the more
score it will add. For example `flac` add `5` score, while `opus` only
add `3`.

If there are two `flac` audio, their score will be a tie. But the
previous version, the smallest index (the index that is checked first)
will win.