feat(mpv): cycle order now start from current track in cycle-track
The first track in the cycle order is always the current track, followed
by the remaining tracks in sequence.

Previously, the cycle order always started from track 1, regardless of
the current track. For example, if the current track was 3, the order
would be 3, 1, 2, 4, 5. Now, the order continues from the current track,
so it will be 3, 4, 5, 1, 2 instead.